Fix float handling in properties dialog and grid normalization
When editing several selected texts with differing sizes, the value
typed in the dialog came back as a String and was only converted back
when the original was an Integer: since the sizes became float, the
string ended up replacing the text content. Add the symmetric Float
branch. Parse PointG coordinates with Float.parseFloat in
DialogParameters. In normalizeCoordinates() the round-up-to-grid idiom
(raw+step-1)/step*step only worked with integer division; with floats
it translated the whole drawing off-grid, so use Math.ceil instead.

+++ src/fidocadj/dialogs/DialogParameters.java
@@ -445,10 +445,10 @@ public final class DialogParameters extends JDialog
                             break;
                         }
                         if (pd.parameter instanceof PointG) {
-                            ((PointG) pd.parameter).x = Integer
-                                    .parseInt(jtf[tc++].getText());
-                            ((PointG) pd.parameter).y = Integer
-                                    .parseInt(jtf[tc++].getText());
+                            ((PointG) pd.parameter).x = Float
+                                    .parseFloat(jtf[tc++].getText());
+                            ((PointG) pd.parameter).y = Float
+                                    .parseFloat(jtf[tc++].getText());
                         } else if (pd.parameter instanceof String) {
                             pd.parameter = jtf[tc++].getText();
                         } else if (pd.parameter instanceof Boolean) {
